requirements.txt
fastapi
uvicorn
streamlit
requests
sqlalchemy
google-generativeai
python-dotenv
pydantic
